requirements.txt
# Core dependencies
fastapi
uvicorn
streamlit
requests
apscheduler
feedparser
python-dotenv

# Optional dependencies for enhanced features
# Uncomment to enable OpenAI GPT integration:
openai>=1.0.0

# Development dependencies (optional)
# pytest
# black
# flake8